% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blue_cascade.R
\docType{data}
\name{blue_cascade}
\alias{blue_cascade}
\title{Council blue menu background and text colors}
\format{
A named list of 8 lists. Each item contains a named list
of two items:
- "background" is the background or fill color
- "color" is the text or contrast color
\describe{
\item{level1}{Level 1, contains the darkest background and lightest color}
\item{level2}{Level 2}
\item{level3}{Level 3}
\item{level4}{Level 4}
\item{level5}{Level 5}
\item{level6}{Level 6}
\item{level7}{Level 7}
\item{level8}{Level 8, contains the lightest background and darkest color}
}
}
\source{
\url{https://metrocouncil.org}
}
\usage{
blue_cascade
}
\description{
A named list of background and text colors for cascading menus
}
\note{
See these colors in action in \code{vignette("Color")}
}
\examples{
blue_cascade$level1$background
blue_cascade$level2$color
}
\seealso{
Other aesthetics:
\code{\link{colors}},
\code{\link[=council.pal]{council.pal()}},
\code{\link[=council_layout]{council_layout()}},
\code{\link[=council_pal2]{council_pal2()}},
\code{\link[=map_council_continuous]{map_council_continuous()}},
\code{\link[=scale_color_council]{scale_color_council()}},
\code{\link[=scale_fill_council]{scale_fill_council()}},
\code{\link[=theme_council]{theme_council()}}
}
\concept{aesthetics}
\keyword{datasets}
