% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_council.R
\name{council.pal}
\alias{council.pal}
\title{Council colors as palettes}
\usage{
council.pal(n, name)
}
\arguments{
\item{n}{numeric, value between 1-8}

\item{name}{character, starting color name. One of \code{councilR::colors}.}
}
\description{
\code{scale_color_council()} and \code{scale_fill_council} give
a color palette starting with \code{councilBlue}, followed by \code{cdGreen} and \code{mtsRed},
and then a couple other Council colors which will look
decent afterwards (up to 8 levels).

Adapted from \code{\link[RColorBrewer:brewer.pal]{RColorBrewer::brewer.pal()}}
}
\examples{
\dontrun{
library(scales)
library(councilR)
show_col(council_pal2())
}
}
\seealso{
Other aesthetics:
\code{\link{blue_cascade}},
\code{\link{colors}},
\code{\link[=council_layout]{council_layout()}},
\code{\link[=council_pal2]{council_pal2()}},
\code{\link[=map_council_continuous]{map_council_continuous()}},
\code{\link[=scale_color_council]{scale_color_council()}},
\code{\link[=scale_fill_council]{scale_fill_council()}},
\code{\link[=theme_council]{theme_council()}}
}
\concept{aesthetics}
