% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_council.R
\name{council_pal2}
\alias{council_pal2}
\title{Discrete Council palettes}
\usage{
council_pal2()
}
\value{
council.pal
}
\description{
Adapted from \code{\link[scales:pal_brewer]{scales::pal_brewer()}}.
Stripped down all arguments, just to show the core
}
\seealso{
Other aesthetics:
\code{\link{blue_cascade}},
\code{\link{colors}},
\code{\link[=council.pal]{council.pal()}},
\code{\link[=council_layout]{council_layout()}},
\code{\link[=map_council_continuous]{map_council_continuous()}},
\code{\link[=scale_color_council]{scale_color_council()}},
\code{\link[=scale_fill_council]{scale_fill_council()}},
\code{\link[=theme_council]{theme_council()}}
}
\concept{aesthetics}
