% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_council_theme.R
\name{council_theme}
\alias{council_theme}
\title{DEPRECATED Council ggplot2 theme}
\usage{
council_theme(
  use_showtext = FALSE,
  size_header = 22,
  size_axis_title = 14,
  size_legend_title = 14,
  size_axis_text = 11,
  size_legend_text = 10,
  size_caption = 8,
  size_margin = 10
)
}
\arguments{
\item{use_showtext}{Logical, whether to use Council fonts.}

\item{size_header}{Header font size in pt. Default is 22}

\item{size_axis_title}{Axis title font size in pt. Default is 14.}

\item{size_legend_title}{Legend title font size in pt. Default is 14.}

\item{size_axis_text}{Axis text font size in pt. Default is 11}

\item{size_legend_text}{Legend test font size in pt. Default is 10.}

\item{size_caption}{Caption font size in pt. Default is 8.}

\item{size_margin}{Margin size in pt. Default is 10}
}
\value{
an error
}
\description{
DEPRECATED Council ggplot2 theme
}
