% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_read_table}
\alias{download_read_table}
\title{Download and read in an Excel file or CSV from a remote URL}
\usage{
download_read_table(url, exdir, force_download = FALSE, ...)
}
\arguments{
\item{url}{character, file location URL. Must end with either ".xlsx", ".xls", or ".csv"}

\item{exdir}{character, directory location where to save downloaded document}

\item{force_download}{logical, whether to force a fresh download, regardless
of whether the file exists already. Default value is \code{FALSE}.}

\item{...}{Additional arguments passed to readxl::read_excel() or readr::read_csv}
}
\value{
tibble
}
\description{
Download and read in an Excel file or CSV from a remote URL
}
\examples{

download_read_table("https://www.dot.state.mn.us/traffic/data/reports/Current_CC_StationList.xlsx",
"_transportation/data-raw/mndot/",
sheet = 1)


}
