% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_counties.R
\name{fetch_counties}
\alias{fetch_counties}
\title{Fetch standardized county geography}
\usage{
fetch_counties(core = TRUE, ...)
}
\arguments{
\item{core}{logical, whether to include all counties in the MPO.}

\item{...}{Arguments passed to \verb{[tigris::counties]}}
}
\value{
An \verb{[sf]} object containing the
}
\description{
Fetch standardized county geography
}
\examples{
\dontrun{
fetch_counties()
}
}
\seealso{
Other spatial helpers: 
\code{\link{snippets_install}()}
}
\concept{spatial helpers}
