% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_county_geo.R
\name{fetch_county_geo}
\alias{fetch_county_geo}
\alias{fetch_ctu_geo}
\title{Fetch standardized geographies}
\usage{
fetch_county_geo(core = TRUE, ...)

fetch_ctu_geo(core = TRUE, ...)
}
\arguments{
\item{core}{logical, whether to include all counties in the MPO.
Default is \code{TRUE}.}

\item{...}{Arguments passed to \verb{[tigris]} functions}
}
\value{
An \code{\link[sf:sf]{sf}} object containing specified geographies.
}
\description{
The default \code{fetch_county_geo()} to return county outlines, plus a suite of other functions to return more niche geographies.

To get city, township, and unorganized territory (CTU) boundaries, use \code{fetch_ctu_geo()}.
}
\examples{
\dontrun{
library(ggplot2)

fetch_county_geo() \%>\%
  ggplot() +
  geom_sf() +
  theme_void()

fetch_ctu_geo() \%>\%
  ggplot() +
  geom_sf(fill = "grey90") +
  theme_void() +
  geom_sf_text(aes(label = CTU_NAME),
    colour = "black",
    check_overlap = F,
    size = 2
  )
}

}
\seealso{
Other spatial helpers:
\code{\link[=snippets_install]{snippets_install()}}
}
\concept{spatial helpers}
