% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_FRED_Oracle.R
\name{FRED_oracle_connection}
\alias{FRED_oracle_connection}
\alias{fred_oracle_connection}
\alias{import_from_FRED_oracle}
\alias{import_from_fred_oracle}
\title{FRED-Oracle}
\usage{
FRED_oracle_connection(
  uid = keyring::key_get("FREDOracle.uid"),
  pwd = keyring::key_get("FREDOracle.pwd"),
  url = keyring::key_get("FREDOracle.url"),
  dsn = keyring::key_get("FREDOracle.dsn")
)

fred_oracle_connection(
  uid = keyring::key_get("FREDOracle.uid"),
  pwd = keyring::key_get("FREDOracle.pwd"),
  url = keyring::key_get("FREDOracle.url"),
  dsn = keyring::key_get("FREDOracle.dsn")
)

import_from_FRED_oracle(
  table_name,
  uid = keyring::key_get("FREDOracle.uid"),
  pwd = keyring::key_get("FREDOracle.pwd"),
  url = keyring::key_get("FREDOracle.url"),
  dsn = keyring::key_get("FREDOracle.dsn")
)

import_from_fred_oracle(
  table_name,
  uid = keyring::key_get("FREDOracle.uid"),
  pwd = keyring::key_get("FREDOracle.pwd"),
  url = keyring::key_get("FREDOracle.url"),
  dsn = keyring::key_get("FREDOracle.dsn")
)
}
\arguments{
\item{uid}{character, FRED-Oracle user id. Default value is
\code{keyring::key_get("FREDOracle.uid")},}

\item{pwd}{character, FRED-Oracle password. Default value is
\code{keyring::key_get("FREDOracle.pwd")}.}

\item{url}{character, FRED-Oracle URL. Default value is
\code{keyring::key_get("FREDOracle.url")}.}

\item{dsn}{character, FRED-Oracle data source name. Default value is
\code{keyring::key_get("FREDOracle.dsn")}.}

\item{table_name}{character, which table to pull. This should include the
database name, separated by a period (i.e. "RESEARCH_WEB.TABLE_NAME")}
}
\value{
\code{fred_oracle_connection()} - An ODBC or JDBC connection object

\code{import_from_fred_oracle()} - Requested table
}
\description{
Functions for interacting with FRED-Oracle.
}
\details{
Both "FRED" and "fred" capitalization are supported.
\itemize{
\item \code{fred_oracle_connection()} creates an connection to FRED-Oracle.
This function will \emph{not} automatically close the connection, so take
care to use \code{\link[DBI:dbDisconnect]{DBI::dbDisconnect()}} once you are done.
\item \code{import_from_fred_oracle()} imports a given table from FRED-Oracle. The connection will
be automatically closed after the table is imported.
}
}
\note{
See \code{vignette("Credentials")} to review credential management.

FRED-Oracle requires its own user id and password, separate from your Met Council
credentials. Contact Matt Schroeder in CD-Research for assistance.

You must be set up with the appropriate database drivers to use these functions.

\strong{Windows} users need ODBC with Oracle drivers. You must also have \code{FREDOracle.dsn} set in your keyring.
Contact IS support for ODBC and driver installation.

\strong{Mac} users need \code{unixodbc} and \code{freetds}, plus Java and JDBC drivers. Additionally,
you must have set \code{FREDOracle.url} in your keyring and \code{JDBC_HOME} in your .Renviron.
See instructions in the
\href{http://mtapshiny1p/MT/Strategic_Initiatives/Onboarding/rodbc.html#configuring-oracle-connections---use-rjdbc-on-any-platform}{onboarding guide}.

Further examples can be found in \code{vignette("Databases")}.
}
\examples{
\dontrun{
library(councilR)
library(DBI)

# create connection
conn <- fred_oracle_connection()

# pull table using SQL
DBI::dbGetQuery(conn, "SELECT * FROM RESEARCH_WEB.RES_PERMIT_TYPE")

# disconnect
DBI::dbDisconnect(conn)

# import a specific table, with no additional SQL logic
import_from_FRED(table_name = "RESEARCH_WEB.RES_PERMIT_TYPE")
}

}
\seealso{
Other database functions:
\code{\link[=FRED_connection]{FRED_connection()}},
\code{\link[=emissions_connection]{emissions_connection()}},
\code{\link[=gis_connection]{gis_connection()}}
}
\concept{database functions}
