% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_FRED.R
\name{FRED_connection}
\alias{FRED_connection}
\alias{fred_connection}
\alias{import_from_FRED}
\alias{import_from_fred}
\title{FRED}
\usage{
FRED_connection(
  uid = keyring::key_get("councilR.uid"),
  pwd = keyring::key_get("councilR.pwd"),
  db = "CD_RESEARCH_WEB",
  prod = TRUE
)

fred_connection(
  uid = keyring::key_get("councilR.uid"),
  pwd = keyring::key_get("councilR.pwd"),
  db = "CD_RESEARCH_WEB",
  prod = TRUE
)

import_from_FRED(
  table_name,
  uid = keyring::key_get("councilR.uid"),
  pwd = keyring::key_get("councilR.pwd"),
  db = "CD_RESEARCH_WEB",
  prod = TRUE
)

import_from_fred(
  table_name,
  uid = keyring::key_get("councilR.uid"),
  pwd = keyring::key_get("councilR.pwd"),
  db = "CD_RESEARCH_WEB",
  prod = TRUE
)
}
\arguments{
\item{uid}{character, your network ID.
Default is \code{keyring::key_get("councilR.uid")}.}

\item{pwd}{character, your network password.
Default is \code{keyring::key_get("councilR.pwd")}.}

\item{db}{character, database name. Default is \code{"CD_RESEARCH_WEB"}.}

\item{prod}{logical, whether to pull from the test or production db.
Default is \code{TRUE}.}

\item{table_name}{character, which table to pull.}
}
\value{
\code{fred_connection()} - A S4 Microsoft SQL Server object

\code{import_from_FRED()} - Requested table
}
\description{
Functions for interacting with FRED, the main CD Research database, stored
Azure SQL Server
}
\details{
Both "FRED" and "fred" capitalization are supported.
\itemize{
\item \code{fred_connection()} creates an S4 Microsoft SQL Server object for FRED.
This function will \emph{not} automatically close the connection, so take
care to use \code{\link[DBI:dbDisconnect]{DBI::dbDisconnect()}} once you are done.
\item \code{import_from_fred()} imports a given table from FRED. The connection will
be automatically closed after the table is imported.
}
}
\note{
See \code{vignette("Credentials")} to review credential management.

You must be set up with the appropriate database drivers to use these functions.

\strong{Windows} users need ODBC with Microsoft SQL. Contact IS support for ODBC installation.

\strong{Mac} users need \code{unixodbc}, \code{freetds}, and properly configured \code{odbc.ini}.

See instructions in the
\href{http://mtapshiny1p/MT/Strategic_Initiatives/Onboarding/rodbc.html}{onboarding guide}.
Further examples can be found in \code{vignette("Databases")}.
}
\examples{
\dontrun{
library(councilR)
library(DBI)

# set credentials if you haven't already
keyring::key_set_with_value("councilR.uid", "mc\\\\myuuid")
keyring::key_set_with_value("councilR.pwd", "password")

# create connection
conn <- FRED_connection(prod = FALSE)

# pull table using SQL
DBI::dbGetQuery(conn, "SELECT * FROM GQ_UNIT WHERE UNIT_ZIP = 55104")

# disconnect
DBI::dbDisconnect(conn)

# import a specific table, with no additional SQL logic
import_from_FRED(table_name = "GQ_UNIT", prod = FALSE)
}

}
\seealso{
Other database functions:
\code{\link[=FRED_oracle_connection]{FRED_oracle_connection()}},
\code{\link[=emissions_connection]{emissions_connection()}},
\code{\link[=gis_connection]{gis_connection()}}
}
\concept{database functions}
