% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_council.R
\name{map_council_continuous}
\alias{map_council_continuous}
\title{Council map starters}
\usage{
map_council_continuous(
  df,
  .fill,
  .lwd = 0.5,
  .low = "#8c510a",
  .mid = "white",
  .high = "#01665e",
  .midpoint = 0
)
}
\arguments{
\item{df}{The \code{sf} object to be mapped}

\item{.fill}{For continuous data, the variable (column name) which
should be used to fill polygons}

\item{.lwd}{numeric, line width of polygons, default setting is \code{0.5}.}

\item{.low}{character, color name or hex code to fill the lowest number.
Default is \code{"#8c510a"} (brown)}

\item{.mid}{character, color name or hex code to fill the midpoint number.
Default is \code{"white"}}

\item{.high}{character,  color name or hex code to fill the highest number.
Default is \code{"#01665e"} (teal)}

\item{.midpoint}{numeric, midpoint of a diverging color scheme.
Default is \code{0}.}
}
\value{
a \link[ggplot2:ggplot2]{ggplot2} plot
}
\description{
Building on functionality for creating maps when mapping continuous data, \code{map_council_continuous()} could save some work.
}
\examples{
\dontrun{
library(ggplot2)
library(councilR)

fetch_ctu_geo() \%>\%
  map_council_continuous(.fill = ALAND, .midpoint = 5e7)
}

}
\seealso{
Other aesthetics:
\code{\link{blue_cascade}},
\code{\link{colors}},
\code{\link[=council.pal]{council.pal()}},
\code{\link[=council_layout]{council_layout()}},
\code{\link[=council_pal2]{council_pal2()}},
\code{\link[=scale_color_council]{scale_color_council()}},
\code{\link[=scale_fill_council]{scale_fill_council()}},
\code{\link[=theme_council]{theme_council()}}
}
\concept{aesthetics}
