% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mn_fips_codes.R
\docType{data}
\name{mn_fips_codes}
\alias{mn_fips_codes}
\title{Minnesota county FIPS codes}
\format{
A data.frame with 5 columns and 87 rows
\describe{
\item{state}{character, state abbreviation}
\item{state_code}{character, two digit state code}
\item{state_name}{character, full state name}
\item{county_code}{character, three digit county code}
\item{county}{character, full county name}
}
}
\source{
\code{{tigris}} dataset, \code{fips_codes}.
}
\usage{
mn_fips_codes
}
\description{
A data.frame with county codes and names for the
state of Minnesota. Derived from \code{{tigris}} fips_codes.
}
\examples{
mn_fips_codes
}
\concept{data}
\keyword{datasets}
