% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_council.R
\name{scale_fill_council}
\alias{scale_fill_council}
\title{Fill scale}
\usage{
scale_fill_council(..., aesthetics = "fill")
}
\arguments{
\item{...}{Arguments passed \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}}

\item{aesthetics}{The names of the aesthetics that this scale works with.}
}
\value{
a Scale object
}
\description{
Fill scale
}
\examples{
\dontrun{
test <- tibble(x = c(1:8), y = c(1:8), col = c(1:8))
ggplot(test, aes(x = x, y = y, col = as.factor(col), fill = as.factor(col))) +
  geom_point(size = 4, pch = 21) +
  scale_fill_council()
}
}
\seealso{
Other aesthetics:
\code{\link{blue_cascade}},
\code{\link{colors}},
\code{\link[=council.pal]{council.pal()}},
\code{\link[=council_layout]{council_layout()}},
\code{\link[=council_pal2]{council_pal2()}},
\code{\link[=map_council_continuous]{map_council_continuous()}},
\code{\link[=scale_color_council]{scale_color_council()}},
\code{\link[=theme_council]{theme_council()}}
}
\concept{aesthetics}
