% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippets.R
\name{snippets_install}
\alias{snippets_install}
\alias{snippets}
\title{Install councilR snippets}
\usage{
snippets_install(install_path = NULL, update = TRUE)
}
\arguments{
\item{install_path}{character, where should the snippets be installed? If \code{NULL},
the snippets will install to a default path based on the current version of
RStudio.}

\item{update}{logical, should existing snippets be updated in place if there are any
conflicts? Default is yes (\code{TRUE}). Otherwise, new snippets are appended to
the end of the existing file, ensuring that you can recover your previous
snippets by editing the snippets file.}
}
\description{
This function installs a set of R snippets that
are helpful when working with Metropolitan Council projects in in
RStudio. By default, the snippets are installed where RStudio will find
them. If you haven't previously installed snippets to RStudio, these
snippets will mask some of the built-in snippets that ship with RStudio.
}
\details{
Snippets will only appear when writing in the snippet's language.
For example, HTML snippets will not appear while editing a ".R" file, but will
appear when editing a ".html" file.
}
\note{
This function is adapted from  \href{https://github.com/gadenbuie/js4shiny/blob/main/R/snippets.R}{\code{js4shiny::snippets_install()}}
}
\section{Updating Existing Snippets}{
 If you already have snippets installed,
you can  you can have the installed snippets update the existing snippets
in place with \code{update = TRUE}. Or you can append the new snippets to the
existing snippets files with \code{update = FALSE}. This option is desirable if
you want to make sure that no snippets are overwritten. The newer snippets
will mask older snippets, but no data will be lost.
}

\examples{
\dontrun{
snip_tmp <- tempfile("snippets")
dir.create(snip_tmp)
snippets_install(snip_tmp)
}

}
\seealso{
Other spatial helpers:
\code{\link[=fetch_county_geo]{fetch_county_geo()}}
}
\author{
Garrick Aden-Buie \email{g.adenbuie@gmail.com},
Liz Roten \email{liz.roten@metc.state.mn.us}
}
\concept{spatial helpers}
