% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_council.R
\name{theme_council}
\alias{theme_council}
\alias{theme_council_open}
\alias{theme_council_geo}
\title{Council ggplot2 theme}
\usage{
theme_council(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  use_showtext = FALSE,
  use_manual_font_sizes = FALSE,
  font_sizes = list(title = 22, subtitle = 16, axis_title = 14, axis_text = 11,
    legend_title = 14, legend_text = 10, caption = 8, strip = 14)
)

theme_council_open(...)

theme_council_geo(...)
}
\arguments{
\item{base_size}{numeric, base font size, given in pts. Default is \code{11}}

\item{base_family}{character, base font family. Default is \code{""}}

\item{base_line_size}{numeric, base size for line elements. Default is \code{base_size/22}}

\item{base_rect_size}{numeric, base size for \code{rect} elements. Default is \code{base_size/22}}

\item{use_showtext}{logical, whether to use Council fonts.}

\item{use_manual_font_sizes}{logical, use supplied font sizes.}

\item{font_sizes}{named list, font sizes. Only used if \code{use_manual_font_sizes} is \code{TRUE}}

\item{...}{arguments passed to \code{theme_council()}}
}
\value{
a \code{\link[ggplot2:theme]{ggplot2::theme()}} object
}
\description{
The default \code{theme_council()} plus a more simple \code{theme_council_open()} for making MetCouncil figures. \code{theme_council()} will be appropriate in most cases while \code{theme_council_open()} is appropriate for single scatter plots or line graphs. For geospatial plots, \code{theme_council_geo()} may be useful to set some initial parameters.

Please note that the y-axis text is horizontal, and long axis names will need to be wrapped;  \code{\link[stringr:str_wrap]{stringr::str_wrap()}} is useful for managing length.
For example, consider using this piece of code: \code{labs(y = stringr::str_wrap("Axis labels are now horizontal, but you still need to insert some code to wrap long labels", width = 15))}
}
\note{
Further examples in \code{vignette("Color")}.
}
\section{Council fonts}{
There are a few requirements needed if you want to use Council font families.
\enumerate{
\item You must have the optional \href{https://github.com/yixuan/showtext}{\code{{showtext}}}
and \href{https://github.com/yixuan/sysfonts}{\code{{sysfonts}}} packages installed.
\item You must have the following fonts installed:
}
\itemize{
\item HelveticaNeueLT Standard Light, Standard Condensed, and Standard Medium Condensed
\item Arial Narrow
\item Palatino Linotype
If you do not have the font files handy, contact a package author or your manager.
}
}

\section{Font size suggestions}{
Generally, font sizes should be no smaller than 11 point.

If you are creating plots for a presentation, you should increase all font sizes.
Check the default font sizes in the presentation platform you are working in
and adjust from there. For example, Microsoft PowerPoint sets slide headings at 77 point and
body text at 46 point. You will likely want to make the plot title less than 77 point so you
can preserve the text hierarchy in the slide.
}

\section{Theme modifications}{
If you wish to make changes to this theme template, such as removing
the legend, adjusting text positions,  access the source code by typing
\code{councilR::theme_council} in the console. You can then copy/paste the
function into whatever document you are working on and make changes there.

If you want to make a change to the default template, open an Issue on GitHub
and ask for the change to be made, and/or make the change yourself and make
a pull request.
}

\examples{
\dontrun{
library(ggplot2)
library(councilR)

ggplot(datasets::iris) +
  aes(Sepal.Length, Sepal.Width, color = Species) +
  geom_point() +
  scale_color_viridis_d() +
  theme_council(
    use_showtext = TRUE,
    use_manual_font_sizes = TRUE
  )

fetch_ctu_geo() \%>\%
  ggplot() +
  geom_sf() +
  theme_council_geo()
}

}
\seealso{
Other aesthetics:
\code{\link{blue_cascade}},
\code{\link{colors}},
\code{\link[=council.pal]{council.pal()}},
\code{\link[=council_layout]{council_layout()}},
\code{\link[=council_pal2]{council_pal2()}},
\code{\link[=map_council_continuous]{map_council_continuous()}},
\code{\link[=scale_color_council]{scale_color_council()}},
\code{\link[=scale_fill_council]{scale_fill_council()}}
}
\concept{aesthetics}
