% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_council.R
\name{theme_council_open}
\alias{theme_council_open}
\title{Appropriate for single scatterplots / line graphs}
\usage{
theme_council_open(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  use_showtext = FALSE,
  use_manual_font_sizes = FALSE,
  font_sizes = list(title = 22, subtitle = 16, axis_title = 14, axis_text = 11,
    legend_title = 14, legend_text = 10, caption = 8, strip = 14)
)
}
\arguments{
\item{base_size}{numeric, base font size, given in pts. Default is \code{11}}

\item{base_family}{character, base font family. Default is \code{""}}

\item{base_line_size}{numeric, base size for line elements. Default is \code{base_size/22}}

\item{base_rect_size}{numeric, base size for rect elements. Default is \code{base_size/22}}

\item{use_showtext}{logical, whether to use Council fonts.}

\item{use_manual_font_sizes}{logical, use supplied font sizes.}

\item{font_sizes}{named list of font sizes. Only used if \code{use_manual_font_sizes} is \code{TRUE}}
}
\value{
a \code{\link[ggplot2:theme]{ggplot2::theme()}} object
}
\description{
Appropriate for single scatterplots / line graphs
}
\examples{
\dontrun{
library(ggplot2)
library(councilR)

ggplot(datasets::iris) +
  aes(Sepal.Length, Sepal.Width, color = Species) +
  geom_point() +
  scale_color_viridis_d() +
  theme_council_open(
    use_showtext = TRUE,
    use_manual_font_sizes = TRUE
  )
}
}
\seealso{
Other aesthetics: 
\code{\link{blue_cascade}},
\code{\link{colors}},
\code{\link{council.pal}()},
\code{\link{council_pal2}()},
\code{\link{council_theme}()},
\code{\link{scale_color_council}()},
\code{\link{scale_fill_council}()},
\code{\link{theme_council}()}
}
\concept{aesthetics}
