% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_github_action_council.R
\name{use_github_action_council}
\alias{use_github_action_council}
\title{Use GitHub Actions template}
\usage{
use_github_action_council(type = "check", path = ".github/workflows")
}
\arguments{
\item{type}{"check". the type of Action to take}
}
\description{
Use GitHub Actions template
}
\examples{
use_github_action_council()

}
