% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{colors}
\alias{colors}
\title{Council colors}
\format{
A list of 29 character objects
\describe{
\item{councilBlue}{Council blue}
\item{cdGreen}{Community Development green}
\item{mtsRed}{Metropolitan Transportation Services red}
\item{esBlue}{Environmental Services blue}
\item{transitBlue}{Metro Transit blue}
\item{transitRed}{Metro Transit red}
\item{transitYellow}{Metro Transit yellow}
\item{suppYellow}{Supplementary yellow}
\item{suppBlack}{Supplmentary black}
\item{suppGray}{Supplementary gray}
\item{suppWhite}{Supplementary white}
\item{metroGreen}{Green Line green}
\item{transitGold}{Gold Line gold}
\item{metroGray}{Lettered line gray}
\item{metroOrange}{Orange Line orange}
\item{playGreen}{Play Features green, a slight variation on \code{cdGreen}}
\item{playBlue}{Play Features blue, a slight variation on \code{councilBlue}}
\item{playLiBlue}{Play Features light blue}
\item{playDaBlue}{Play Features dark blue}
\item{playSalmon}{Play Features salmon, or pink}
\item{playYellow}{Play Features yellow}
\item{metrostatsBlue}{Metro Stats blue}
\item{metrostatsRed}{Metro Stats red}
\item{metrostatsDaPurple}{Metro Stats dark purple}
\item{metrostatsMePurple}{Metro Stats medium purple}
\item{metrostatsLiPurple}{Metro Stats light purple}
\item{metrostatsPink}{Metro Stats pink}
\item{metrostatsTan}{Metro stats tan}
}
}
\source{
\url{https://www.metrotransit.org/Data/Sites/1/media/logos/brand_standards.pdf}
}
\usage{
colors
}
\description{
A named list of Council colors for each division, Metro Transit,
Play Features, and MetroStats in Hex code format
}
\examples{
colors$councilBlue

}
\seealso{
Other aesthetics: 
\code{\link{council_theme}()}
}
\concept{aesthetics}
\keyword{datasets}
