% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/council_layout.R
\name{council_layout}
\alias{council_layout}
\alias{plotly_layout}
\title{Council plotly formatting}
\usage{
council_layout(
  a_plotly,
  main_title = "",
  subtitle = "",
  x_title = "",
  y_title = "",
  legend_title = "",
  ...
)

plotly_layout(
  a_plotly,
  main_title = "",
  subtitle = "",
  x_title = "",
  y_title = "",
  legend_title = "",
  ...
)
}
\arguments{
\item{a_plotly}{\code{\link[plotly:plot_ly]{plotly::plot_ly()}} object}

\item{main_title}{character, plot title}

\item{subtitle}{character, plot subtitle}

\item{x_title}{character, x-axis title}

\item{y_title}{character y-axis title}

\item{legend_title}{character, legend title}

\item{...}{additional parameters passed to \code{\link[plotly:layout]{plotly::layout()}}}
}
\value{
\code{\link[plotly:plot_ly]{plotly::plot_ly()}} object with Council styling
}
\description{
Council plotly formatting
}
\note{
Further examples in \code{vignette("Color")}.

The \code{subtitle} returned is an annotation, and so cannot be further modified.
If you want to modify the subtitle aesthetics, leave \code{subtitle = ""} and
add a subtitle using \code{plotly::layout(annotations = list(...))}.
}
\examples{
\dontrun{

library(plotly)
library(councilR)

plotly::plot_ly(
  type = "scatter",
  mode = "markers",
  data = iris,
  x = ~Sepal.Length,
  y = ~Sepal.Width,
  color = ~Species,
  hoverinfo = "text",
  hovertext = ~ paste0(
    Species, "<br>",
    "Sepal Length: ", Sepal.Length, "<br>",
    "Sepal Width: ", Sepal.Width
  ),
  marker = list(
    size = 10,
    opacity = 0.8
  )
) \%>\%
  plotly_layout(
    main_title = "Iris",
    subtitle = "Sepal characteristics",
    x_title = "Sepal Length",
    y_title = "Sepal Width",
    legend_title = "Species",
    legend = list(orientation = "h")
  )
}
}
\seealso{
Other aesthetics: 
\code{\link{blue_cascade}},
\code{\link{colors}},
\code{\link{council.pal}()},
\code{\link{council_pal2}()},
\code{\link{map_council_continuous}()},
\code{\link{scale_color_council}()},
\code{\link{scale_fill_council}()},
\code{\link{theme_council}()}
}
\concept{aesthetics}
