% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_council.R
\name{council_pal2}
\alias{council_pal2}
\title{Discrete Council palettes}
\usage{
council_pal2()
}
\value{
council.pal
}
\description{
Modified from \code{\link[scales:brewer_pal]{scales::brewer_pal()}}.
Stripped down all arguments, just to show the core
}
\seealso{
Other aesthetics: 
\code{\link{blue_cascade}},
\code{\link{colors}},
\code{\link{council.pal}()},
\code{\link{scale_color_council}()},
\code{\link{scale_fill_council}()},
\code{\link{theme_council}()}
}
\concept{aesthetics}
