% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/council_theme.R
\name{council_theme}
\alias{council_theme}
\title{Council ggplot2 theme}
\usage{
council_theme(
  use_showtext = FALSE,
  size_header = 22,
  size_axis_title = 14,
  size_legend_title = 14,
  size_axis_text = 11,
  size_legend_text = 10,
  size_caption = 8,
  size_margin = 10
)
}
\arguments{
\item{use_showtext}{logical, whether to use Council fonts.}

\item{size_header}{header font size in px}

\item{size_axis_title}{axis title font size in px}

\item{size_legend_title}{legend title font size in px}

\item{size_axis_text}{axis text font size in px}

\item{size_legend_text}{legend test font size in px}

\item{size_caption}{caption font size in px}

\item{size_margin}{the margin size in pt}
}
\value{
a \code{\link[ggplot2:theme]{ggplot2::theme()}} object
}
\description{
Council ggplot2 theme
}
\section{Council fonts}{
If you wish to use Council fonts, you must have the following fonts installed:
\itemize{
\item HelveticaNeueLT Standard Light, Standard Condensed, and Standard Medium Condensed
\item Arial Narrow
\item Palatino Linotype
}

You must also have the \href{https://github.com/yixuan/showtext}{showtext} and \href{https://github.com/yixuan/sysfonts}{sysfonts}
packages installed.
}

\section{Font size suggestions}{
If you are creating plots for a presentation, you should increase all font sizes.
Check the default font sizes in the presentation platform you are working in
and adjust from there.
}

\section{Theme modifications}{
If you wish to make changes to this theme template, such as removing
the legend, adusting text positons,  access the source code by typing
\code{councilR::council_theme} in the console. You can then copy/paste the
function into whatever document you are working on and make changes.
}

\seealso{
Other aesthetics: 
\code{\link{colors}}
}
\concept{aesthetics}
