% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/councilR-package.R
\docType{package}
\name{councilR-package}
\alias{councilR}
\alias{councilR-package}
\title{councilR: Functions and Templates for the Metropolitan Council}
\description{
A collection of commonly used templates, color palettes, and functions for R users at the Metropolitan Council.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/Metropolitan-Council/councilR/issues}
}

}
\author{
\strong{Maintainer}: Liz Roten \email{liz.roten@metc.state.mn.us} (\href{https://orcid.org/0000-0002-5346-3549}{ORCID})

Other contributors:
\itemize{
  \item Metropolitan Council [copyright holder]
  \item Garrick Aden-Buie \email{garrick@adenbuie.com} (\href{https://orcid.org/0000-0002-7111-0077}{ORCID}) [contributor]
  \item Ellen Esch \email{ellen.esch@metc.state.mn.us} (\href{https://orcid.org/0000-0002-4253-0910}{ORCID}) [contributor]
  \item Mauricio León \email{mauricio.leon@metc.state.mn.us} [contributor]
}

}
\keyword{internal}
