% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_emissions.R
\name{emissions_connection}
\alias{emissions_connection}
\alias{import_from_emissions}
\title{Import data table from greenhouse gas emissions scenario planning database}
\usage{
emissions_connection(
  uid = keyring::key_get("councilR.uid"),
  pwd = keyring::key_get("councilR.pwd"),
  local = TRUE,
  db = "CD_Emissions"
)

import_from_emissions(
  table_name,
  uid = keyring::key_get("councilR.uid"),
  pwd = keyring::key_get("councilR.pwd"),
  local = TRUE,
  db = "CD_Emissions"
)
}
\arguments{
\item{uid}{character, your network ID.
Default is \code{keyring::key_get("councilR.uid")}.}

\item{pwd}{character, your network password.
Default is \code{keyring::key_get("councilR.pwd")}.}

\item{local}{logical, whether to pull from the onsite database or Azure.
Default is \code{TRUE}.}

\item{db}{character, database name. Default is \code{"CD_Emissions"}.}

\item{table_name}{character, which table to pull.}
}
\value{
Requested table

\code{import_from_emissions()} - Requested table
}
\description{
Import data table from greenhouse gas emissions scenario planning database
}
\details{
To access Azure, your IP address must be approved. Contact Sean Molloy.
}
\note{
See \code{vignette("Credentials")} to review credential management.
You must be set up with the appropriate database drivers to use these functions.
\strong{Windows} users need ODBC with Microsoft SQL. Contact IS support for ODBC installation.
\strong{Mac} users need \code{unixodbc} and \code{freetds}. See instructions in the
\href{https://furry-adventure-596f3adb.pages.github.io/database-connections.html}{onboarding guide}
}
\examples{
\dontrun{

library(councilR)
library(DBI)

# set credentials if you haven't already
keyring::key_set_with_value("councilR.uid", "mc\\\\myuuid")
keyring::key_set_with_value("councilR.pwd", "password")

# create connection
conn <- emissions_connect()

# pull table using SQL
DBI::dbGetQuery(conn, "SELECT * FROM metro_energy.vw_electricity_residential_ctu")

# disconnect
DBI::dbDisconnect(conn)

# get specific tables
t_electricity_residential_ctu <- import_from_emissions(
  table_name = "metro_energy.vw_electricity_residential_ctu"
)

t_eia_energy_consumption_state <- import_from_emissions(
  table_name = "state_energy.eia_energy_consumption_state"
)

t_utility_natural_gas_by_ctu <- import_from_emissions(
  table_name = "metro_energy.vw_utility_natural_gas_by_ctu"
)
}

}
\seealso{
Other database functions: 
\code{\link{FRED_connection}()},
\code{\link{gis_connection}()}
}
\concept{database functions}
