% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_county_geo.R
\name{fetch_county_geo}
\alias{fetch_county_geo}
\title{Fetch standardized county geography}
\usage{
fetch_county_geo(core = TRUE, ...)
}
\arguments{
\item{core}{logical, whether to include all counties in the MPO. Default is \code{TRUE}.}

\item{...}{Arguments passed to \verb{[tigris::counties]}}
}
\value{
An \code{\link{sf}} object containing county geographies.
}
\description{
Fetch standardized county geography
}
\note{
This function relies on \verb{[\{rlang\}]} internal functions.
}
\examples{
\dontrun{
fetch_county_geo()
}

}
\seealso{
Other spatial helpers: 
\code{\link{snippets_install}()}
}
\concept{spatial helpers}
