% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_emissions.R
\name{import_from_emissions}
\alias{import_from_emissions}
\title{Import data table from greenhouse gas emissions scenario planning database}
\usage{
import_from_emissions(
  table_name,
  uid = getOption("councilR.uid"),
  pwd = getOption("councilR.pwd"),
  local = TRUE,
  db = "CD_Emissions"
)
}
\arguments{
\item{table_name}{character, which table to pull.}

\item{uid}{character, your network ID.
Default is \code{getOption("councilR.uid")}. If you are accessing Azure,
this value should be your full email address.}

\item{pwd}{character, your network password.
Default is \code{getOption("councilR.pwd")}. For example, \code{"my_password"}}

\item{local}{logical, whether to pull from the onsite database or Azure.
Default is \code{TRUE}.}

\item{db}{character, database name. Default is \code{"CD_Emissions"}.}
}
\value{
Requested table
}
\description{
Import data table from greenhouse gas emissions scenario planning database
}
\details{
To access Azure, your IP address must be approved. Contact Sean Molloy.
}
\note{
See \code{vignette("Options")} to review package options.
You must be set up with the appropriate database drivers to use this function.
\strong{Windows} users need ODBC with Microsoft SQL. Contact IS support for ODBC installation.
\strong{Mac} users need \code{unixodbc} and \code{freetds}. See instructions in the
\href{https://furry-adventure-596f3adb.pages.github.io/database-connections.html}{onboarding guide}

\if{html}{\out{<div class="sourceCode">}}\preformatted{This function relies on `[\{rlang\}]` internal functions.
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
library(councilR)

# set options if you haven't already
options(
  councilR.uid = "mc\\\\you",
  councilR.pwd = "mypwd"
)

t_electricity_residential_ctu <- import_from_emissions(
  table_name = "metro_energy.vw_electricity_residential_ctu"
)
t_eia_energy_consumption_state <- import_from_emissions(
  table_name = "state_energy.eia_energy_consumption_state"
)
  table_name = "metro_energy.vw_utility_natural_gas_by_ctu"
  )
}
}
