% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_emissions.R
\name{import_from_emissions}
\alias{import_from_emissions}
\title{Import data table from Emissions Database}
\usage{
import_from_emissions(
  uid = councilR.uid,
  pwd = councilR.pwd,
  table_name = "metro_energy.vw_electricity_residential_ctu",
  local = TRUE,
  serv = "dbsqlcl11t.test.local,65414",
  db = "CD_Emissions"
)
}
\arguments{
\item{uid}{character, your network id.
Default is \code{getOption("councilR.uid")}. For example, \code{"mc\\\\rotenle"}}

\item{pwd}{character, your network password.
Default is \code{getOption("councilR.pwd")}. For example, \code{"my_password"}}

\item{table_name}{character, which table to pull}

\item{serv}{character, database server.
Default is \code{"dbsqlcl11t.test.local,65414"} (the test database).}

\item{db}{character, database name. Default is \code{"CD_Emissions"}}
}
\value{
a table from the CD_Emissions database.
}
\description{
WARNING: Function error may results in RStudio crash. Requires a password for access to the database.
}
\note{
To make connection seamless, add \code{councilR.uid} and \code{councilR.pwd} to your \code{.Rprofile}. Edit your \code{.Rprofile} as such\preformatted{options(
    ...,
    councilR.uid = "mc\\\\myuid",
    councilR.pwd = keyring::key_get("MetC")
)
}

See more details in \href{https://github.com/Metropolitan-Council/councilR/blob/main/vignettes/Options.Rmd}{\code{{councilR}} documentation}.
}
\examples{
\dontrun{
library(councilR)

# set options if you haven't already
options(
  councilR.uid = "mc\\\\myuid",
  councilR.pwd = "mypwd"
)

t_electricity_residential_ctu <- import_from_emissions(table_name = "metro_energy.vw_electricity_residential_ctu")
t_eia_energy_consumption_state <- import_from_emissions(table_name = "state_energy.eia_energy_consumption_state")
t_utility_natural_gas_by_ctu <- import_from_emissions(table_name = "metro_energy.vw_utility_natural_gas_by_ctu")
}
}
