% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_FRED.R
\name{import_from_FRED}
\alias{import_from_FRED}
\alias{import_from_fred}
\title{Import data table from FRED (main CD Research database) data stored
Azure SQL Server}
\usage{
import_from_FRED(
  table_name,
  uid = getOption("councilR.uid"),
  pwd = getOption("councilR.pwd"),
  db = "CD_RESEARCH_WEB",
  prod = TRUE
)

import_from_fred(
  table_name,
  uid = getOption("councilR.uid"),
  pwd = getOption("councilR.pwd"),
  db = "CD_RESEARCH_WEB",
  prod = TRUE
)
}
\arguments{
\item{table_name}{character, which table to pull.}

\item{uid}{character, your network ID.
Default is \code{getOption("councilR.uid")}.}

\item{pwd}{character, your network password.
Default is \code{getOption("councilR.pwd")}. For example, \code{"mypwd"}}

\item{db}{character, database name. Default is \code{"CD_RESEARCH_WEB"}.}

\item{prod}{logical, whether to pull from the test or production db.
Default is \code{TRUE}.}
}
\value{
Requested table
}
\description{
Import data table from FRED (main CD Research database) data stored
Azure SQL Server
}
\note{
See \code{vignette("Options")} to review package options.
You must be set up with the appropriate database drivers to use this function.
\strong{Windows} users need ODBC with Microsoft SQL. Contact IS support for ODBC installation.
\strong{Mac} users need \code{unixodbc}, \code{freetds}, and properly configured \code{odbc.ini}.
See instructions in the
\href{https://furry-adventure-596f3adb.pages.github.io/database-connections.html}{onboarding guide}
and contact package maintainer for assistance.

\if{html}{\out{<div class="sourceCode">}}\preformatted{This function relies on `[\{rlang\}]` internal functions.
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
library(councilR)

# set options if you haven't already
options(
  councilR.uid = "mc\\\\you",
  councilR.pwd = "mypwd"
)

import_from_FRED(table_name = "GQ_UNIT", prod = FALSE)
}
}
