% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_gis.R
\name{import_from_gis}
\alias{import_from_gis}
\title{Import a dataset from ArcCatalog}
\usage{
import_from_gis(
  query,
  dbname = "GISLibrary",
  uid = getOption("councilR.uid"),
  pwd = getOption("councilR.pwd")
)
}
\arguments{
\item{query}{character, string with the database connection and feature class}

\item{dbname}{character, database name. Default is \code{"GISLibrary"}.
Must match the database name in \code{query}.}

\item{uid}{character, user ID. Default is \code{getOption("councilR.uid")}}

\item{pwd}{character, user password. Default is \code{getOption("councilR.pwd")}.}
}
\value{
an \url{sf} object
}
\description{
Import a dataset from ArcCatalog
}
\note{
See \code{vignette("Options")} to review package options.
You must be set up with the appropriate database drivers to use this function.
\strong{Windows} users need ODBC with Microsoft SQL. Contact IS support for ODBC installation.
\strong{Mac} users need \code{unixodbc} and \code{freetds}. See instructions in the \href{https://furry-adventure-596f3adb.pages.github.io/database-connections.html}{onboarding guide}

\if{html}{\out{<div class="sourceCode">}}\preformatted{ This function relies on `[\{rlang\}]` internal functions.
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
library(councilR)

options(councilR.uid = "mc\\\\uid", councilR.pwd = "pwd")

# query db name matches
import_from_gis(query = "GISLibrary.dbo.AIRPORTS", dbname = "GISLibrary")
}
}
