% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_gis.R
\name{import_from_gis}
\alias{import_from_gis}
\title{Import a dataset from ArcCatalog}
\usage{
import_from_gis(
  query,
  dbname = "GISLibrary",
  uid = getOption("councilR.uid"),
  pwd = getOption("councilR.pwd")
)
}
\arguments{
\item{query}{character, string with the database connection and feature class}

\item{dbname}{character, database name. Usually either \code{"GIS"} or \code{"GISLibrary"}}

\item{uid}{character, user ID. default is \code{getOption("councilR.uid")}}

\item{pwd}{character, user password. Default is \code{getOption("councilR.pwd")}.}
}
\value{
an \code{sf} object
}
\description{
Import a dataset from ArcCatalog
}
\note{
See \code{vignette("Options")} to review package options.
}
\examples{
\dontrun{
import_from_gis("GISLibrary.dbo.AIRPORTS")
}
}
