% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_gpkg.R
\name{import_from_gpkg}
\alias{import_from_gpkg}
\alias{import_gpkg}
\title{DEPRECATED Import an \code{sf} object using a geopackage URL}
\usage{
import_from_gpkg(
  link,
  save_file = FALSE,
  save_path = getwd(),
  .crs = 4326,
  keep_temp = FALSE,
  .quiet = TRUE,
  ...
)

import_gpkg(
  link,
  save_file = FALSE,
  save_path = getwd(),
  .crs = 4326,
  keep_temp = FALSE,
  .quiet = TRUE,
  ...
)
}
\arguments{
\item{link}{character, URL for the \code{.gpkg} object}

\item{save_file}{logical, whether to save the downloaded file.
File will be saved in .RDS format}

\item{save_path}{character, the path where the downloaded file should be saved.
Default location is the working directory.}

\item{.crs}{numeric, the CRS code or string used to transform the download.
Default is \code{4326} (WGS 84)}

\item{keep_temp}{character, whether to keep the temporary download.
Default is \code{FALSE}.}

\item{.quiet}{logical, suppress messages. Default is \code{TRUE}.}

\item{...}{additional parameters passed to \link[sf:st_read]{sf::read_sf}}
}
\value{
\code{\link[sf:sf]{sf::sf()}} object
}
\description{
Deprecated.
This function is particularly useful for importing data from
\href{https://gisdata.mn.gov/}{Minnesota Geospatial Commons}
when access to GISLibrary is unavailable.
}
\examples{
\dontrun{

library(councilR)

# import regional parks from Minnesota Geospatial Commons
import_from_gpkg("https://resources.gisdata.mn.gov/pub/gdrs/data/pub/us_mn_state_metc/plan_parks_regional/gpkg_plan_parks_regional.zip")
# import the "RegionalEnvironmentalJusticeByCensusTract" layer only
import_gpkg("https://resources.gisdata.mn.gov/pub/gdrs/data/pub/us_mn_state_metc/trans_tpp2050/gpkg_trans_tpp2050.zip", layer = "RegionalEnvironmentalJusticeByCensusTract")
}
}
