% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_council.R
\name{map_council_continuous}
\alias{map_council_continuous}
\title{Council map starters}
\usage{
map_council_continuous(
  df,
  .fill,
  .lwd = 0.5,
  .low = "#8c510a",
  .mid = "white",
  .high = "#01665e",
  .midpoint = 0
)
}
\arguments{
\item{df}{The \code{sf} object to be mapped}

\item{.fill}{For continuous data, the variable which should be used to fill polygons}

\item{.lwd}{The line width of polygons, default setting is \code{0.5}.}

\item{.low}{The color to fill the lowest number. Default is brown.}

\item{.mid}{The color to fill the midpoint number. Default is white}

\item{.high}{The color to fill the highest number. Default is teal}

\item{.midpoint}{The midpoint of a diverging color scheme. Default is \code{0}.}
}
\value{
a \link{ggplot2} plot
}
\description{
Building on functionality for creating maps When mapping continuous data, \code{map_council_continuous()} could save some work.
}
\note{
This function relies on \verb{[\{rlang\}]} internal functions.
}
\examples{
\dontrun{
library(ggplot2)
library(councilR)

fetch_ctu_geo() \%>\%
  map_council_continuous(.fill = ALAND, .midpoint = 5e7)
}

}
\seealso{
Other aesthetics: 
\code{\link{blue_cascade}},
\code{\link{colors}},
\code{\link{council.pal}()},
\code{\link{council_pal2}()},
\code{\link{scale_color_council}()},
\code{\link{scale_fill_council}()},
\code{\link{theme_council}()}
}
\concept{aesthetics}
