% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_council.R
\name{theme_council}
\alias{theme_council}
\title{Council ggplot2 theme}
\usage{
theme_council(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  use_showtext = FALSE,
  use_manual_font_sizes = FALSE,
  font_sizes = list(title = 22, subtitle = 16, axis_title = 14, axis_text = 11,
    legend_title = 14, legend_text = 10, caption = 8, strip = 14)
)
}
\arguments{
\item{base_size}{numeric, base font size, given in pts. Default is \code{11}}

\item{base_family}{character, base font family. Default is \code{""}}

\item{base_line_size}{numeric, base size for line elements. Default is \code{base_size/22}}

\item{base_rect_size}{numeric, base size for rect elements. Default is \code{base_size/22}}

\item{use_showtext}{logical, whether to use Council fonts.}

\item{use_manual_font_sizes}{logical, use supplied font sizes.}

\item{font_sizes}{named list of font sizes. Only used if \code{use_manual_font_sizes} is \code{TRUE}}
}
\value{
a \code{\link[ggplot2:theme]{ggplot2::theme()}} object
}
\description{
Council ggplot2 theme
}
\section{Council fonts}{
There are a few requirements needed if you want to use Council font families.
\enumerate{
\item You must have the optional \href{https://github.com/yixuan/showtext}{\code{{showtext}}}
and \href{https://github.com/yixuan/sysfonts}{\code{{sysfonts}}} packages installed.
\item You must have the following fonts installed:
}
\itemize{
\item HelveticaNeueLT Standard Light, Standard Condensed, and Standard Medium Condensed
\item Arial Narrow
\item Palatino Linotype
If you do not have the font files handy, contact a package author or your manager.
}
\enumerate{
\item The font family and file names much match those listed below
}
\itemize{
\item "HelveticaNeueLT Std Cn" = "HelveticaNeueLTStd-Cn.otf"
\item "HelveticaNeueLT Std Lt" = "HelveticaNeueLTStd-Lt.otf"
\item "Arial Narrow" ="ARIALN.TTF"
\item "HelveticaNeueLT Std Med Cn" = "HelveticaNeueLTStd-MdCn.otf")
\item "Palatino Linotype" = "pala.ttf"
}
\enumerate{
\item Consider options for your OS
}
\itemize{
\item On Windows, be sure that the fonts are installed for the entire system, not just a single user.
These are located in \verb{C:/Windows/Fonts}.
\item On Mac OSX, be sure that fonts are installed for the user. These are located in \verb{~/Library/Fonts}.
Consider copying the entire system font directory (\verb{/Library/Fonts}) to the user font directory for easy access.
}
}

\section{Font size suggestions}{
Generally, font sizes should be no smaller than 11 point.

If you are creating plots for a presentation, you should increase all font sizes.
Check the default font sizes in the presentation platform you are working in
and adjust from there. For example, Microsoft PowerPoint sets slide headings at 77 point and
body text at 46 point. You will likely want to make the plot title less than 77 point so you
can preserve the text hierarchy in the slide.
}

\section{Theme modifications}{
If you wish to make changes to this theme template, such as removing
the legend, adjusting text positions,  access the source code by typing
\code{councilR::theme_council} in the console. You can then copy/paste the
function into whatever document you are working on and make changes there.

If you want to make a change to the default template, open an Issue on GitHub
and ask for the change to be made, and/or make the change yourself and make
a pull request.
}

\examples{
\dontrun{
library(ggplot2)
library(councilR)

ggplot(datasets::iris) +
  aes(Sepal.Length, Sepal.Width, color = Species) +
  geom_point() +
  scale_color_viridis_d() +
  theme_council(
    use_showtext = TRUE,
    use_manual_font_sizes = TRUE
  )
}

}
\seealso{
Other aesthetics: 
\code{\link{blue_cascade}},
\code{\link{colors}},
\code{\link{council_theme}()}
}
\concept{aesthetics}
